Moar Balls, Moar Functionality
Added a BP_MessengerComponent (badly named) which I can add to a scene instance along with a BoxCollider and implement functionality on an already existing scene instance! Sick! Using this to send a signal (observer pattern (maybe a retarded way to do it *shrugs in ASCII*)) to the BP_GameController which then calls 'RemoveBall' which doesnt do anything except print a string for now but Ill make it return the ball to the pool tmrw